+++ gt7dashboard/tabs/racelines_tab.py
@@ -18,7 +18,7 @@ from bokeh.palettes import Plasma11 as palette
 from bokeh.plotting import figure
 
 from gt7dashboard.gt7lap import Lap
-from .base_tab import GT7Tab
+from .GT7Tab import GT7Tab
 from gt7dashboard.gt7helper import car_name
 from gt7dashboard.gt7help import get_help_div
 
@@ -394,7 +394,3 @@ class RaceLinesTab(GT7Tab):
         """Initialize the race lines tab"""
         # Update available laps
         self.update_lap_options()
-
-    def get_tab_panel(self):
-        """Create a TabPanel for this tab"""
-        return TabPanel(child=self.layout, title="Race Lines")
